Creating a visual layout test for pherrit-links
Creates a script (built & executed by webpack) which creates an HTML
file that shows all the possible variants of the pherrit links.

The purpose of this file is to work on the styling in isolation while
having an overall view of all the variants at once. This compiled file
is not committed to the repo.